ändring i css, footer, flyttade searchbar i searchpage och tog bort searchbarjsx och serchbarcss
