--- src/components/SearchBar.css
@@ -1,77 +1,0 @@
-.overlay {
-  position: fixed;
-  top: -10rem;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.searchContainer {
-  position: relative;
-  padding: 20px;
-  background-color: white;
-  border-radius: 5px;
-  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.2);
-  width: 80%;
-  max-width: 600px;
-  text-align: center;
-}
-
-.searchForm {
-  display: flex;
-  align-items: center;
-}
-
-.searchInput {
-  width: 100%;
-  padding: 10px;
-  border: 1px solid #ddd;
-  border-radius: 5px;
-  font-size: 16px;
-}
-
-.dropdown {
-  position: absolute;
-  top: 100%;
-  left: 0;
-  width: 100%;
-  background-color: white;
-  border: 1px solid #ddd;
-  border-radius: 5px;
-  list-style: none;
-  padding: 0;
-  margin-top: 5px;
-  z-index: 10;
-}
-
-.dropdownItem {
-  padding: 10px;
-  cursor: pointer;
-  border-bottom: 1px solid #ddd;
-}
-
-.dropdownItem.selected {
-  background-color: #f0f0f0;
-}
-
-.search-button {
-  background: none;
-  border: none;
-  cursor: pointer;
-  color: black;
-  font-size: 24px;
-}
-
-.search-button:hover {
-  color: #e68900;
-}
-
-
-@media (min-width: 768px) {
-        .searchContainer {
-        width: 35%;
-        }
-}
