+++ src/index.css
@@ -134,6 +134,87 @@ ul {
   background: #f5f5f5;
 }
 
+/* ------- searchbar-------- */
+.overlay {
+  /* position: fixed;
+  top: -10rem;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center; */
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center ;
+  z-index: 900; 
+}
+
+.searchContainer {
+  position: relative;
+  padding: 20px;
+  background-color: white;
+  border-radius: 5px;
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.2);
+  width: 80%;
+  max-width: 600px;
+  text-align: center;
+}
+
+.searchForm {
+  display: flex;
+  align-items: center;
+}
+
+.searchInput {
+  width: 100%;
+  padding: 10px;
+  border: 1px solid #ddd;
+  border-radius: 5px;
+  font-size: 16px;
+}
+
+.dropdown {
+  position: absolute;
+  top: 100%;
+  left: 0;
+  width: 100%;
+  background-color: white;
+  border: 1px solid #ddd;
+  border-radius: 5px;
+  list-style: none;
+  padding: 0;
+  margin-top: 5px;
+  z-index: 10;
+}
+
+.dropdownItem {
+  padding: 10px;
+  cursor: pointer;
+  border-bottom: 1px solid #ddd;
+}
+
+.dropdownItem.selected {
+  background-color: #f0f0f0;
+}
+
+.search-button {
+  background: none;
+  border: none;
+  cursor: pointer;
+  color: black;
+  font-size: 24px;
+}
+
+.search-button:hover {
+  color: #e68900;
+}
+
+
+
+
 .text-capitalize {
   text-transform: capitalize;
 }
@@ -534,7 +615,9 @@ h2 {
   padding: 30px 20px;
   text-align: center;
   margin-top: auto;
-  display: none;
+  display: none; 
+  /* width: 100%; */
+
 }
 
 .footer-container {
@@ -721,6 +804,11 @@ ul.breadcrumb li a:hover {
     bottom: auto;
   }
 
+  .searchContainer {
+    width: 35%;
+    }
+
+
   .footer {
     display: block;
     margin-top: 1rem;
@@ -758,7 +846,7 @@ ul.breadcrumb li a:hover {
     padding-top: 4rem;
  }
 
- .favorites-container {SS
+ .favorites-container {
   padding-top: 6rem;
 }
